Adds tagging to hamilton
Gets at #4. Tags are meant to navigate nodes. Currently its TBD
how exactly we'll use them, but they've suggested value in a lot
of places. Tags are specifically set to strings, meant for marking/
indexing variables.

The user will use them by looking for Variables with tags. As the
fields are optional, this will be fully backwards compatible.